+++ services/src/functional-tests/src/test/groovy/org/openkilda/functionaltests/spec/switches/SwitchRulesSpec.groovy
@@ -123,6 +123,8 @@ class SwitchRulesSpec extends BaseSpecification {
                 sw.ofVersion == "OF_12" && data.installRulesAction != InstallRulesAction.INSTALL_BROADCAST)
         assumeFalse("Unable to run the test because only NoviFlow switches support installation of BFD catch rule",
                 !sw.noviflow && data.installRulesAction == InstallRulesAction.INSTALL_BFD_CATCH)
+        assumeFalse("Unable to run the test because only NoviFlow switches support installation of Round Trip Latency "
+                + "rule", !sw.noviflow && data.installRulesAction == InstallRulesAction.INSTALL_ROUND_TRIP_LATENCY)
 
         given: "A switch without any rules"
         def defaultRules = northbound.getSwitchRules(sw.dpId).flowEntries
@@ -165,6 +167,10 @@ class SwitchRulesSpec extends BaseSpecification {
                         [
                                 installRulesAction: InstallRulesAction.INSTALL_BFD_CATCH,
                                 cookie            : Cookie.CATCH_BFD_RULE_COOKIE
+                        ],
+                        [
+                                installRulesAction: InstallRulesAction.INSTALL_ROUND_TRIP_LATENCY,
+                                cookie            : Cookie.ROUND_TRIP_LATENCY_RULE_COOKIE
                         ]
                 ],
                 uniqueSwitches