+++ services/src/kilda-core/kilda-persistence-neo4j/src/main/java/org/openkilda/persistence/repositories/impl/Neo4jRepositoryFactory.java
@@ -1,4 +1,4 @@
-/* Copyright 2018 Telstra Open Source
+/* Copyright 2019 Telstra Open Source
  *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
@@ -24,6 +24,7 @@ import org.openkilda.persistence.repositories.FlowPairRepository;
 import org.openkilda.persistence.repositories.FlowPathRepository;
 import org.openkilda.persistence.repositories.FlowRepository;
 import org.openkilda.persistence.repositories.IslRepository;
+import org.openkilda.persistence.repositories.KildaConfigurationRepository;
 import org.openkilda.persistence.repositories.LinkPropsRepository;
 import org.openkilda.persistence.repositories.RepositoryFactory;
 import org.openkilda.persistence.repositories.SwitchRepository;
@@ -125,5 +126,10 @@ public class Neo4jRepositoryFactory implements RepositoryFactory {
     @Override
     public BfdSessionRepository createBfdSessionRepository() {
         return new Neo4JBfdSessionRepository(sessionFactory, transactionManager);
+    }
+
+    @Override
+    public KildaConfigurationRepository createKildaConfigurationRepository() {
+        return new Neo4jKildaConfigurationRepository(sessionFactory, transactionManager);
     }
 }