+++ services/src/messaging/src/main/java/org/openkilda/messaging/model/FlowDto.java
@@ -151,6 +151,9 @@ public class FlowDto implements Serializable {
     @JsonProperty("priority")
     private Integer priority;
 
+    @JsonProperty("pinned")
+    private boolean pinned;
+
     public FlowDto() {
     }
 
@@ -178,6 +181,7 @@ public class FlowDto implements Serializable {
      * @param state             flow state
      * @param maxLatency        max latency
      * @param priority          flow priority
+     * @param pinned            pinned flag
      */
     @JsonCreator
     @Builder(toBuilder = true)
@@ -201,7 +205,8 @@ public class FlowDto implements Serializable {
                    @JsonProperty(Utils.FLOW_PATH) final PathInfoData flowPath,
                    @JsonProperty("state") FlowState state,
                    @JsonProperty("max_latency") Integer maxLatency,
-                   @JsonProperty("priority") Integer priority) {
+                   @JsonProperty("priority") Integer priority,
+                   @JsonProperty("pinned") boolean pinned) {
         this.flowId = flowId;
         this.bandwidth = bandwidth;
         this.ignoreBandwidth = ignoreBandwidth;
@@ -223,6 +228,7 @@ public class FlowDto implements Serializable {
         this.state = state;
         this.maxLatency = maxLatency;
         this.priority = priority;
+        this.pinned = pinned;
     }
 
     /**
@@ -238,13 +244,14 @@ public class FlowDto implements Serializable {
      * @param destinationSwitch destination switch
      * @param destinationPort   destination port
      * @param destinationVlan   destination vlan id
+     * @param pinned            pinned flag
      */
     public FlowDto(String flowId,
                    long bandwidth,
                    boolean ignoreBandwidth,
                    String description,
                    SwitchId sourceSwitch, int sourcePort, int sourceVlan,
-                   SwitchId destinationSwitch, int destinationPort, int destinationVlan) {
+                   SwitchId destinationSwitch, int destinationPort, int destinationVlan, boolean pinned) {
         this(flowId,
                 bandwidth,
                 ignoreBandwidth,
@@ -259,7 +266,7 @@ public class FlowDto implements Serializable {
                 destinationPort,
                 sourceVlan,
                 destinationVlan,
-                null, 0, null, null, null, null);
+                null, 0, null, null, null, null, pinned);
     }
 
     public FlowDto(FlowPayload input) {
@@ -279,7 +286,8 @@ public class FlowDto implements Serializable {
                 input.getDestination().getVlanId(),
                 null, 0, null, null,
                 input.getMaxLatency(),
-                input.getPriority());
+                input.getPriority(),
+                input.isPinned());
     }
 
     @JsonIgnore
@@ -341,6 +349,18 @@ public class FlowDto implements Serializable {
     public boolean containsSwitchInPath(SwitchId switchId) {
         return flowPath.getPath().stream()
                 .anyMatch(node -> node.getSwitchId().equals(switchId));
+    }
+
+    /**
+     * Checks creation params to figure out whether they are aligned or not.
+      * @return validation result
+     */
+    @JsonIgnore
+    public boolean isValid() {
+        if (isAllocateProtectedPath() && isPinned()) {
+            return false;
+        }
+        return true;
     }
 
     /**